Tried to use musl as a translation layer on host.
The idea is to build musl on host and change all symbols
with "objcopy --prefix-symbols=<prefix>", so that they are different
from the symbols in the host libc. For applications that use the
translation layer, we build a static library with all object files
first and use "objcopy --redefine-syms=<file>" to change all symbols we
have in musl to use the prefixed version. In theory, all functions that
should be translated to M³ use the prefixed version, whereas everything
else just refers to the host libc.

Although being hacky, this approach basically works. For example,
basename seems to work exactly as it should. However, it's quite easy
to miss that a symbol accidentally refers to the host libc and thus the
approach is quite error prone. Also, all tools that use fts_* currently
still refer to the host libc, because we also need to use the same for
our bsdutils library that we do for musl. And finally, to get C++
applications running, we also need to do the translation in the
libstdc++. That's even more annoying, because it's a library provided
by the host system, where it's unclear whether:
1. it exists at all (the static version)
2. it is compatible with musl

Overall I'm not convinced that this approach is worth the trouble.