resmng: optionally verify hash before application start.
Boot scripts can not specify the expected hash for a specific
application that should be started via the "hash" attribute in
the "app" tag. The resource manager will then ask the executable file
for its hash and check it against the expected hash.

Additionally, there is a new open flag called HASH, which requests that
the file system obtains a hash for the opened file and returns it upon
the GET_HASH request to the client. Thus, the resource manager uses
this flag to open the executable file if a hash was specified in the
boot script.

Currently, no file supports hashing, but a secure wrapper around m3fs
will have this support in the near future. It is also possible to
extend root and support this check for boot modules.