boot: added ability to specify environment variables in boot scripts.
This is intended for scripts that require certain settings like a
specific file system, a specific gem5 config, etc.. Previously, it was
pretty annoying that users had to remember the required settings for
specific boot scripts, which typically resulted in multiple attempts to
figure out how a script is supposed to be run.

Therefore, these scripts now specify the corresponding environment
variables in an "env" tag below "config". Note that a warning is issued
in case the user has also set an environment variable and the values
differ.