Auto-initialize tile ids from environment.
Now that we have the tile ids in the environment, we should pass them to
all applications (in case someone wants to decode a PMP EP for
example). And it also allows us to auto-initialize the ids from the
environment on first use. This avoids potential initialization problems
and makes it simpler for bare-metal components.