Overhauled FPGA usage concept.
Previously it was required to have a copy of the platform/hw repo on the
FPGA PC and we always used the Python-Code, Scripts, and Bitfiles from
this repo. This was quite annoying, because we already have this repo
as a submodule in the M³ repo and they of course have to be in sync.

I've changed that now so that everything is done from the local
platform/hw repo and we simply copy the relevant files to the FPGA PC
beforehand. That is, we copy the Python-Code before starting something
on the FPGA. Additionally, I've added a command to b called loadfpga to
load a specified Bitfile onto the FPGA. In this case, the Bitfile is
copied to the FPGA PC before loading it.

For that reason, the environment variables for the hw platform changed.
The new ones are:

M3_HW_FPGA_HOST: SSH alias for the FPGA PC
M3_HW_FPGA_DIR: directory on the FPGA PC, relative to the home
  directory, that is used for temporary files.
M3_HW_FPGA_NO: FPGA number (number to add to base IP 192.168.42.240)
M3_HW_VIVADO: path to the vivado/vivado-lab binary on the FPGA PC

All of them need to be explicitly set in order to avoid that we
accidentally connect to the wrong FPGA just because the user forgot to
set a variable.