Make FetchIRQ a privileged command.
Previously, the TCU offered a dedicated register to read the currently
handled IRQ and acknowledge it. However, now that we allow 32-bit
accesses to TCU registers, acknowledging becomes tricky if we don't have
bits that indicated "now the register content is complete".

Since we already have this semantic for privileged commands, I decided
to introduce a new privileged command for that purpose. On x86_64 we
need to know the interrupt number as well, so that the FetchIRQ command
does not only acknowledge the most recent interrupt, but also returns
its number.