Removed support for ARMv7.
We can no longer run it on gem5 for quite some time now, because gem5
removed support for it in favor of ARMv8, apparently. Since we will not
use it on the hardware platform for the forseeable future and we now
have another 32-bit architecture with riscv32, there is no longer any
point for us in supporting it.