kernel: cache multiplexer type.
Now that we are about to make use of it in the kernel ourself, we should
no longer ask the multiplexer for its type everytime (this is not going
to ever change anyway), but do that once at the beginning and cache the
result.

Unfortunately, it's a bit more tricky during initialization because we
cannot make use of threads until root has been started. Since we need to
load TileMux on the tile for root anyway, we solve this problem by
simply hardcoding the multiplexer type instead of asking.