Added request_ep sidecall for accelerator multiplexers.
The idea is that due to size constraints, accelerators have a simple
endpoint-based interface. The multiplexer next to the accelerator
therefore needs to know which endpoints to use and therefore needs to
have a say in which endpoints are actually allocated.

So far, endpoints are allocated from the kernel without the multiplexer
knowing about it. That's fine for TileMux and Linux as they don't care
which endpoints are uses, but only about how many each activity can use
which can be enforced via quota by the kernel.

I've therefore introduced another multiplexer type for accelerators and
a new sidecall to handle that. That is, for accelerator multiplexers the
kernel will ask the multiplexer via the request_ep sidecall which
endpoint should be used.