riscv32: start execution at address 0.
The problem is that the core in the hardware platform has its code
memory at 0x0 and its data memory at 0x10000, so that we put the
environment at address 0x10000. Since we also want to support larger
applications on other RV32 cores, we put the code behind the
environment. However, as we need the same reset vector, we use 0 for
that in general. The normal baremetal programs (all except straccmux)
use a simple trampoline at address 0 to jump to the actual code.