Make memory offset and env address tile-dependent.
The problem is that the RV32 core next to the accelerators has a
different memory offset than the RV64 cores in the same system. Since
some software components (e.g., kernel, root, etc.) need to load other
components, they need to know these properties for a different tile than
the one they are running on. For that reason, this properties can no
longer be compile-time-known, but need to be determined at runtime
depending on the tile we are talking about.

For that reason, TileDesc now provides access to these properties.
Unfortunately, the PhysAddr type now also needs to know for which tile
the physical address belongs and not only during construction, but also
later when, for example, extracting the EP id. This in turn means that
we cannot (that easily) use PhysAddr like a simple integer (by
implementing all the typical integer operations). As this is not really
needed right now, I've decided to just remove this feature.